Assignment after reseting indentation
